src/main/java/com/howard/leetcode/binary/tree/Solution236.java
package com.howard.leetcode.binary.tree;


/**
 * 二叉树的最近公共祖先
 *
 * @author howard he
 * @create 2019-03-11 10:50
 * @see LowestCommonAncestorSolution
 */
public class Solution236 {

    public TreeNode lowestCommonAncestor(TreeNode root, TreeNode p, TreeNode q) {
        if (root == null) {
            return null;
        }
        if (root == p || root == q) {
            return root;
        }
        TreeNode left = lowestCommonAncestor(root.left, p, q);
        TreeNode right = lowestCommonAncestor(root.right, p, q);
        if (left == null) {
            return right;
        }
        if (right == null) {
            return left;
        }
        return root;
    }
}
